// src/generated/swagger/post-it.ts
export interface LoginResponse {
  account?: Account;
}

export interface RegisterResponse {
  account?: Account;
}

export interface GetAllPostsResponse {
  postsSummaries?: PostSummary[];
}

export interface GetPostResponse {
  post?: Post;
}

export interface CreatePostResponse {
  post?: Post;
}

export interface EditPostResponse {
  post?: Post;
}

export interface Account {
  token?: string;
  author?: string;
}

export interface Post {
  /** @format int64 */
  id?: number;
  title?: string;
  content?: string;
  image?: string;

  /** @format int64 */
  creationTime?: number;
  author?: string;
}

export interface PostSummary {
  /** @format int64 */
  id?: number;
  title?: string;
  author?: string;
}

export interface LoginPayload {
  /** @format email */
  email: string;
  password: string;
}

export interface RegisterPayload {
  /** @format email */
  email: string;
  password: string;
  author: string;
}

export interface GetAllPostsParams {
  /**
   * current page
   * @min 0
   */
  "page-number"?: number;

  /**
   * Page size
   * @min 1
   * @max 50
   */
  "page-size"?: number;

  /**
   * filter using from id
   * @format int64
   */
  "from-id"?: number;
}

export interface CreatePostPayload {
  post: Post;
}

export interface EditPostPayload {
  post?: Post;
}

export namespace User {
  /**
   * No description
   * @tags user
   * @name Login
   * @summary login if the email exists
   * @request POST:/user/login
   */
  export namespace Login {
    export type RequestParams = {};
    export type RequestQuery = {};
    export type RequestBody = LoginPayload;
    export type RequestHeaders = {};
    export type ResponseBody = LoginResponse;
  }
  /**
   * No description
   * @tags user
   * @name Register
   * @summary create new user
   * @request POST:/user/register
   */
  export namespace Register {
    export type RequestParams = {};
    export type RequestQuery = {};
    export type RequestBody = RegisterPayload;
    export type RequestHeaders = {};
    export type ResponseBody = RegisterResponse;
  }
}